Avoid GCM Birthday barrier, AD to ensure target type.
AES-GCM has an issue with random nonce reuse, it is not safe beyond 4b uses.
Because HTTP traffic can be high and secret rotation lax, it's not nonzero that
we get there. Work around this by using the key+random nonce to derive a new
key, that is also bound to the AD.

Also use the AD to attach an encryption context to things. Set this to our
current magic and the session name, to avoid any potential vector of fuzzing
with cookie names to try and deserialize one data set into a different struct.