Drop the exist flag on Get
It didn't really add any value. A usable session object is always returned, and
if the user wants initialization we have a interface for that. Beyond that, I
can't really think of any case where a generic "yes this existed" adds value.
So get rid of it. If the user wants to track new vs. existing, it's easy enough
to add a bool to the session data.