Switch to tink for encryption, support proto serialization
Doing custom stuff for crypto was fun, but tink does all we need (i.e
AES-GCM-SIV) and handles keysets in a way that pushes the user to do better. So
switch to it.

Because we now have proto as a dependency, support it optionally for session
types. A bit more efficient serialization.